Rebases hierarchy
