+++ packages/server/src/async/producers/feed-items/new-measurement-prediction.js
@@ -1,9 +1,7 @@
 const {
   MEASUREMENT_COMPETITOR_TYPE,
 } = require('../../../models/enums/measurement-competitor-type');
-
 const { Producer } = require('../producer');
-
 const { NewMeasurement } = require('./new-measurement');
 
 class NewMeasurementPrediction extends NewMeasurement {
