+++ packages/server/src/async/producers/feed-items/producer-feed-items.js
@@ -1,11 +1,72 @@
 const assert = require('assert');
+const _ = require('lodash');
 
 const { Producer } = require('../producer');
 
 class ProducerFeedItems extends Producer {
 
-  constructor(options) {
-    super(options);
+  constructor(input) {
+    super({});
+    assert(_.isObject(input), 'Input should be an object.');
+    assert(!!_.get(input, 'id'), 'Measurable Id is required.');
+    assert(!!_.get(input, 'agentId'), 'Agent Id is required.');
+    assert(!!_.get(input, 'channelId'), 'Agent Id is required.');
+    this.input = input;
+  }
+
+  /**
+   * @public
+   * @return {Promise<boolean>}
+   */
+  async main() {
+    try {
+      if (await this._isActual() === false) {
+        console.log(this.name, 'Hook is not actual');
+        return true;
+      }
+    } catch (e) {
+      console.error(this.name, e.message, e);
+      return false;
+    }
+
+    try {
+      const agent = await Producer.data.agents.getOne({
+        id: this.input.agentId,
+      });
+      assert(!!_.get(agent, 'id'), 'Agent ID is required.');
+
+      const replacements = await this._getReplacements(agent);
+      const channelId = this.input.channelId;
+      await this._queueFeedItem(replacements, channelId);
+
+      await this._commit();
+      return true;
+    } catch (e) {
+      await this._rollback();
+      console.error(this.name, e.message, e);
+      return false;
+    }
+  }
+
+  /**
+   * @param {Models.Agent} agent
+   * @return {Promise.<{agent: {name: string}}>}
+   * @protected
+   */
+  async _getReplacements(agent) {
+    return {
+      agent: {
+        name: (await _.get(agent, 'name')) || 'Somebody',
+      },
+    }
+  }
+
+  /**
+   * @protected
+   * @return {Promise<boolean>}
+   */
+  async _isActual() {
+    return false;
   }
 
   /**
